refactor: move outlier related classes to lib
